proxy: detect TCP socket hang ups from client or server
We previously `join`ed on piping data from both sides, meaning
that the future didn't complete until **both** sides had disconnected.
Even if the client disconnected, it was possible the server never knew,
and we "leaked" this future.

To fix this, the `join` is replaced with a `select`, so that if either
side hangs up, we end the future.

Closes #434

Signed-off-by: Sean McArthur <sean@seanmonstar.com>